# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Collaboration API
  description: Extended REST API for projects and real-time chat with AI integration
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: https://api.example.com
    description: Production server
  - url: http://localhost:3000
    description: Development server

security:
  - bearerAuth: []

paths:
  /v1/projects:
    get:
      summary: List all projects
      operationId: listProjects
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: sort
          in: query
          schema:
            type: string
            default: createdAt:desc
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      summary: Create a new project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - description
              properties:
                name:
                  type: string
                description:
                  type: string
                metadata:
                  type: object
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'

  /v1/projects/{id}:
    get:
      summary: Get project by ID
      operationId: getProject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    put:
      summary: Update project
      operationId: updateProject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    delete:
      summary: Delete project
      operationId: deleteProject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Project deleted

  /v1/projects/{id}/status:
    patch:
      summary: Update project status
      operationId: updateProjectStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [active, archived, deleted]
      responses:
        '200':
          description: Status updated

  /v1/users:
    get:
      summary: List all users
      operationId: listUsers
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
              properties:
                name:
                  type: string
                email:
                  type: string
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /v1/projects/{id}/members:
    get:
      summary: List project members
      operationId: listProjectMembers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of project members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectMember'
    post:
      summary: Add a member to a project
      operationId: addProjectMember
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - role
              properties:
                user_id:
                  type: string
                role:
                  type: string
                  enum: [editor, viewer]
      responses:
        '201':
          description: Member added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMember'

  /v1/projects/{id}/members/{user_id}:
    put:
      summary: Update a project member's role
      operationId: updateProjectMember
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
              properties:
                role:
                  type: string
                  enum: [editor, viewer]
      responses:
        '200':
          description: Member updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMember'
    delete:
      summary: Remove a member from a project
      operationId: removeProjectMember
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Member removed

  /v1/projects/{id}/analytics:
    get:
      summary: Get project analytics
      operationId: getProjectAnalytics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Analytics'

  /v1/chats:
    get:
      summary: List all chats
      operationId: listChats
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of chats
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Chat'
    post:
      summary: Create a new chat
      operationId: createChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - project_id
                - participants
              properties:
                project_id:
                  type: string
                participants:
                  type: array
                  items:
                    type: string
                metadata:
                  type: object
      responses:
        '201':
          description: Chat created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'

  /v1/chats/{id}:
    get:
      summary: Get chat details
      operationId: getChat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Chat details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
    delete:
      summary: Delete chat
      operationId: deleteChat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Chat deleted

  /v1/chats/{id}/messages:
    get:
      summary: List chat messages
      operationId: listMessages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: before
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
    post:
      summary: Send a message
      operationId: sendMessage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sender_id
                - content
              properties:
                sender_id:
                  type: string
                content:
                  type: string
                type:
                  type: string
                  default: text
                metadata:
                  type: object
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  /v1/chats/{id}/ai/reply:
    post:
      summary: Generate AI response using minimax-m2
      operationId: generateAiReply
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sender_id
              properties:
                sender_id:
                  type: string
                  description: ID of the user requesting AI response
      responses:
        '201':
          description: AI-generated message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Chat not found
        '500':
          description: Failed to generate AI response

  /v1/chats/{id}/ai/think:
    post:
      summary: Generate a sequence of thoughts to solve a problem
      operationId: sequentialThinking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - prompt
              properties:
                prompt:
                  type: string
                  description: The problem to solve.
                goal:
                  type: string
                  description: The desired outcome.
                totalThoughts:
                  type: integer
                  description: The number of thoughts to generate.
                  default: 5
      responses:
        '200':
          description: A sequence of thoughts and a final answer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SequentialThinkingResponse'
        '404':
          description: Chat not found
        '500':
          description: Failed to generate thoughts

  /v1/system/health:
    get:
      summary: Health check
      operationId: healthCheck
      responses:
        '200':
          description: System healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  timestamp:
                    type: string
                    format: date-time

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token for API authentication

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        created_at:
          type: string
          format: date-time

    ProjectMember:
      type: object
      properties:
        user_id:
          type: string
        role:
          type: string
          enum: [owner, editor, viewer]

    Activity:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        action:
          type: string
        timestamp:
          type: string
          format: date-time

    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, archived, deleted]
        members:
          type: array
          items:
            $ref: '#/components/schemas/ProjectMember'
        activity:
          type: array
          items:
            $ref: '#/components/schemas/Activity'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object

    Chat:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
        participants:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        metadata:
          type: object

    Message:
      type: object
      properties:
        id:
          type: string
        chat_id:
          type: string
        sender_id:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [text, image, file]
        created_at:
          type: string
          format: date-time
        metadata:
          type: object
          properties:
            reply_to:
              type: string
            mentions:
              type: array
              items:
                type: string
            model:
              type: string
              description: AI model used (e.g., minimax-m2)
            generated:
              type: boolean
              description: Whether message was AI-generated

    Analytics:
      type: object
      properties:
        project_id:
          type: string
        views:
          type: integer
        stars:
          type: integer
        messages_count:
          type: integer
        last_activity:
          type: string
          format: date-time

    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer

    Thought:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string

    SequentialThinkingResponse:
      type: object
      properties:
        thoughts:
          type: array
          items:
            $ref: '#/components/schemas/Thought'
        final_answer:
          type: string